ARM: [vCPU-ABI] Fix handling of TPIDRURO register
The TPIDRURO register is stored during exception IPC and vCPU
exits. Because it is not part of the normal kernel exception state is
has to be moved to the end of the user-visible exception state.
This fixes a problem with ARM virtualization where an invalid TPIDRURO
might be loaded for a VM.

Change-Id: I0c8ed60e5ae1e54584c23dfd7ec54e1952525e69