ia32: Handle in-kernel faults due to user input
On x86 the user supplies segment registers as part
of the CPU state in exception IPC and vCPU handling.
Giving wrong values will lead to exceptions generated
in ring-0 and make the kernel unhappy.

To handle those situations, we tag all locations where
this can happen and checks against those locations
in case we get an in-kernel exception. If it is
from one of the known locations, we make the
corresponding user-thread fault, and go on.

Currently, I access the five locations with extern symbols.
An improvement could be to put them to an extra section
in assembly and loop over that section in the exception
handling code. However for those five occasions this is
probably not worth it. I also gave them proper names
so that it's more descriptive.

This basically only affects 32bit. However, 64bit
needs to look for wrong ss as well.

With this change we can also further remove checks
in Trap_state::sanitize_user_state and fasten up
vCPU handling.

Change-Id: I2b60d6d4c89937c5afc8d88f0c115a731964ffb0