mips: access VZ timer-related registers with 32bit only
The VCPU extended state defines all timer related registers
as natural-width registers. This looks to be correct according
to the spec, which seems to provide for an extension of the timer
to 64bit in the future. However, at the moment the upper 32bit
are considered reserved on existing 64bit processors. So ignore
them when computing timer values.

Change-Id: I176f0af6d771c632efa1fc2a87c0f1ac96c65f7d