ARM cleanup 'armca9' missuse
There used to be a wide missuse of the armca9 config tag, it was in fact
used to say something like armv7 and newer, but not ARM Cortex-A8.

So I modified the Modules.arm to set armca9 exactly for ARM-Cortex A9
CPUs and for no other CPU and replaced the use with 'armv7 || armv8' and
other extra conditions.

This change also modifies the handling of the 'S' (shareable) bit for
ARMv7+ CPUs, the new model set the 'S' bit for MP configurations and
does not set the 'S' bit for UP configurations. This should catch the
Cortex-A8 case, which should always be compiled as UP.

The change also modifies the handling of the SWP enable bit in ARMv7
SCTLR. The new code checks the MIDR, MPIDR, and ISAR0 for support of SWP
and for multiprocessor extensions and sets the SCTLR bit acording to the
config and the CPU.

This change also contains fixes for the memory attributes used for ARMv7
page tables. The new version uses correct cache and sharable attributes
for ARMv7 (also Cortex A8), and removes cache writeback operations for
Cortex-A8 which also uses cache coherent page-table walks.

Change-Id: I6437195f2122dcafa95939d22eea37f5ab0d07be