Fix race condition in IRQ shortcut
We have to avoid using the IPC/IRQ shortcut when the receiver thread
waits for completion of a DRQ, otherwise the thread might be pushed to
user-level before the DRQ reply is finished, which is a use-after-free
for the DRQ object.

The problem was originally discovered by Christian Ehrhardt
<Christian_Ehrhardt@genua.de>.

This fixes ticket #380.

Change-Id: I19ecc49d7c379e5dd5bf70178bd48d5c08d94ea5