[x86]: Fix platform reset via keyboard controller
On a machine with no PS2 controller Fiasco hangs in the first
while-loop while reading from an IO port. Because reading 0xff from
an IO port is considered an error this change exits the loop and
continues with the reset eventually trying to reset via PCI.

This change also adds the reset-via-PCI method for ia32.

Change-Id: I301db5373910b4d0efd950bd5147b35776f22897