Rework kernel page-table handling
Define the type Kpdir for all architectures to define the type of the
page-table used for the kernel address space. And store a pointer to the
kernel page table in Mem_lyout::kdir (Kmem::kdir).

For most cases Kpdir is just the same as Pdir, however some
architectures or configurations use a separate address space (type) for
the kernel address space (such as ARM with virtualization support).

The generic support for Kmem::kdir for all architectures allows to
translate kernel addresses to physical addresses using
Kmem::kdir->virt_to_phys.

Change-Id: Ibcd50c2ef47a888d4e5780efdc10ffcb41718508