Fix handling of switch_to stack
We have two places where we directly setup a stack frame for a thread
that is then activated via switch_cpu. The first one in when a thread is
prepared for the first time to exit to user land. The second occasion is
the fast return path from IPC that sets up a stack frame to directly
return to user land during IPC.

Both versions need to comply with the stack-pointer alignment and
requirements of the swtich_cpu code and the handlers that are invoked,
the new function Context::prepare_switch_to() provides an abstraction
for that purpose.

This fixes a bug in the AArch64 fast IPC return code.

Change-Id: Ib7cd41b0f85b613a8069787ebd18e73a9d9e6501