ia32: enable hardware controlled performance states (hwp)
Intel introduced hwp to enable the CPU to decide on which
performance states to use. The idea is that the OS gives the
CPU hints, e.g. if it is on AC, it may hint that the CPU can
optimize for performance and when on battery, it may hint the
CPU to optimize for low energy usage. The CPU will then use
these hints and decide autonomously on which frequency core
wattage it is using.

This change enables hwp (if available on the CPU), and
configures it such that the CPU has the most autonomy in
choosing the power states.

Fixes #270

Change-Id: I2ec53a25c67454911605b279c646dfd18966add7