AArch32: Fix processor mode checks for VMs
Implement strict checking of target CPU modes when switching to a VM
with ARM hardware virtualization support. Formerly we did only check
that the target mode is not HYP mode, but ARM CPUs would just stay in
HYP mode if we assign an illegal CPU mode for the VM which is a serious
security problem. With this change we allow only valid modes (User, FIQ,
IRQ, Supervisor, Abort, Undefined, and System) to be set during vCPU
resume, all other modes will be overwritten and set to user mode.

Change-Id: Id75cb1853a98cc7cbb834eacaad61ea137bfd7c1