AArch64: Fix FPU handling in HYP kernel
When we run at EL2 we have to make sure that CPACR_EL1.FPEN is 3 when
the FPU is disabled and we run with HCR.TGE = 1, otherwise we get
undefined instructions instead of FPU traps into EL2.
Also cleaned up the SCTLR_EL1 handling for VMs as we no longer run the
VMM at EL1.

Change-Id: I2159a9096af718195b74b364828c62b43b5ed751