ARM: Fix IPI addressing for clusters
The target list in the interrupt controller (GIC) for IPIs
is a bitmap of 8 bits. With clusters, the physical CPU ID is
a number of (at most) 24 bits, i.e. there's mapping needed
from the physical CPU ID to the target number of an SGI.

As Fiasco allows for arbitrary CPU ordering, the BSP has
to provide a mapping from the physical CPU ID to the
target number. Per default we assume a 1:1 mapping but
also check that proper values are used.

BSPs need to implement the Gic::pcpu_to_sgi() function
if required (i.e. they're a cluster).

Change-Id: I696fb4fb5a5deaa7491f5707e6ef81b7849e89bf