Fix assertion during execute_drq
In particular cross-CPU IRQ triggers in combination with thread
migrations could trigger the assertion in Context::_execute_drq.

The particular condition is that a locally (non-cross CPU) DRQ might be
preemptive as it is in the case of Irq_sender::handle_remote_hit. In
this case the target thread of a DRQ might no longer be on the same CPU
after the DRQ function is executed. In this case we remove the assertion
and just skip the ready-queue handling as this has to be done already in
this case.

Fixes ticket #595.

Change-Id: I0fcb756662167734db709857129803291f89d660