MIPS: fix invalid register allocation for JDB entry
We have to set the result register for the JDB entry as early clobber,
because it is written before the last ASM instruction in the inline ASM
statement.

This fixes broken JDB entries experienced on MIPS64r6 with the IMG
compiler. However, the working cases where just luck with the compilers
register allocation.

Change-Id: Id40d0a703c6c6e3b8a525a3f6bd784a5385571a7