x86: Improve documentation on number of CPUs vs TSS
Care must be taken when increasing the number
of CPUs beyond what's allowed through the config.
For too large configs we exceed the space
available for a globally mapped TSS space and
thus cannot store the 16bit offset to the IO bitmap
anymore. Unfortunately this offset got truncated,
thus leading to faulty IO-port behavior.
This change adds more checks to ensure
sizes are in limit.

Change-Id: I6f8ee81a9f28dd9dd9eb9ec98187bb37b2f0060e