ia32/apic: fixed compiler warning in us_to_apic on 32-bit
scaler_us_to_apic is actually 32-bit. It's set from Cpu::muldiv() which
returns Unsigned32. For amd64 we want to keep scaler_us_to_apic as 64-bit
to be able to directly multiply without the need to load the variable
into a register first (see the amd64 implementation of us_to_apic()).

Change-Id: I444076e391a7f4195493079aba3bb9ebdad4bc3e