IA32: use local APIC ID of current CPU in MSI info
When the user requests an MSI use the APIC-ID of the current CPU in the
MSI address field instead of 0. This fixes problems with systems that
have no CPU with APIC-ID 0 and also might create more local IRQ delivery
when an application uses the same CPU when requesting the MSI
information and handling the IRQ.

This might actually fix ticket #644.

Change-Id: I34e191a14062fe2a9a4b668e74675e1a951476f4