ARM32: load VMPIDR on VM switch and VMM to VM switch
Apply changes from Iecb51f8be3e180707ae223403ff6fb15a2196262 to
ARM32 with slightly different ordering to adapt to the different
alignment constraints.

Since the MPIDR ist not readable from EL0 we use the current MPIDR
value to initialize the vmpidr field. We might mask parts of the
MPIDR which should not be known to the user in a later revision.

Change-Id: I7bc34a4aa2667154cd206455120d44495c556eb9