Added 'U' to ARFLAGS (prevents warnings)
Sometimes binutils is compiled with --enable-deterministic-archives (at
least here with binutils 2.29 on Debian/Sid) which makes are archive
members having zero UIDs, GIDs timestamps, and using consistent file
modes for all files. Setting the timestamps to zero means that the
archive will always be updated as the file in the archive is always
older than the file to be added.

This behavior is IMO only necessary for building distributable static
librares -- which isn't the case here.

The 'U' switch was introduced with binutils 2.23 released in 10/2012.
IMO reasonable old.

This fixes the ar warning about using deterministic mode and always
updating libuart.a.

Change-Id: I499487b74050a0e959a2e269473fd8669ec5918f