IA32: Add option to disable IO page faults
To generate L4 IO-Port page faults we need to decode the instruction
that caused a #GP or #PF exception from the fiasco kernel. To avoid this
we simply send out a normal #GP exception IPC in the case of a
potential IO-Port fault if the new option is enabled. NOTE, the new
option is default on.

Change-Id: Ie026c0288ba1950bff8d04be154162fef36ddd5d
Requires-Change-Id: I67097f6c9d5393c21da40628d523a5a7db09bb63