ARM: Maintain outer cache along with the inner one
Some ARM processors such as Cortex A9 feature an outer cache that needs
to be explicitly maintained in addition to maintaining the common inner
cache hierarchy. In order to remove the burden of dealing with this
cache-architecture detail from userspace code, this change adds code for
carrying out the respective outer cache maintenance operations as part
of the innner cache maintenance.

Change-Id: I171f0f5ae3479bde329e01e17ebbbde8d229116b