Test object pointer before invoking method on it
On some platforms, thread_timer_interrupt() will pass NULL pointer to
Upstream_irq to Timer_tick::handler_all() which in turn forwards it to
Timer_tick::handle_timer(). handle_timer() will attempt to invoke a
non-static method ack() on this NULL object, which is an undefined
behaviour. Without CONFIG_INLINE, the kernel might crash later on in
Upstream_irq::ack because of this. Whether it crashes or not might also
depend on other factors such as compiler version and optimization level.
With the default CONFIG_INLINE though, the invocation of ui->ack() gets
optimized away, so we have been lucky so far.

This fixes #782.

Change-Id: Ia95d274c7c72a388ccd62e5525029c91e1e9d893