Check opcode presence on thread object invocation
Upon invocation, the thread object does not systematically check that
the word with the opcode is present in the UTCB. The check is done for
opcodes that take arguments, but not for opcodes that don't. This commit
therefore adds one unconditional check which is always performed before
any opcode handling.

Fixes ticket #824.

Change-Id: Ib92ad73559ac521190e0039b4cc72d909a34eb19