Allow up to 3328MB of kernel memory for 64bit
We also need to adapt the Syscall_entry code for amd64 syscall
instructions because the code contains a jump instruction that is
limited to +/-2GB from the syscall entry code to the kernel code.

This change moves the dynamically allocated kernel memory out of
this 2GB range from the kernel image and hence would break
syscall for all but the boot CPU. To avoid this we also move the
syscall entry code into a static per-CPU array that is within the
kernel image itself insead part of dynamically allocated per-CPU
memory.

Change-Id: I5341ed9db565c79ad16132306d38409bfc3e3492