ia32/amd64: deterministic CPU bootup
Up until now Fiasco boots CPUs on x86 by issuing the startup IPI via
broadcast. This results in the CPUs booting in random order. For
use-cases where a deterministic CPU assignment of threads to logical
CPUs is required this makes it unnecessarily hard as each time the
mapping of CPU numbers to topology has to be recalculated.

This change makes the CPU numbers deterministic across reboots of the
machine. Fiasco reads the local APIC information from the MADT for each
CPU and reserves a CPU number according to the entry by storing the APIC
ID in an array. When a new CPU boots it looks into the array and
retrieves its reserved CPU number.

This makes the CPU number assignment independent of the order in which
CPUs awake from the startup IPI.

Change-Id: I70e004d4ce334fdc2c8bda76101702f9da4d2203