x86: silence kmem message
The kernel prints a kmem message for each booted CPU with information
about the kernel memory region. Other messages are switched off using an
'if (0)' block.

This change defines an enum 'Print_info' and changes all 'if (0)'
message blocks to use this enum value. The kernel memory region message
is wrapped accordingly.

Change-Id: I310fcbea550988b9c85174814d63fc9043b6708b