ia32/ux: Fix Idt setup for UX
As the Idt setup does not use Vmem_alloc::page_alloc()
we need to call page_map explicitly so that UX gets
the mapping. To be able to do this, we need to
make Vmem_alloc::page_map public. Do this by removing
it from the generic Vmem_alloc class as it is
only available for x86-based platforms anyway, and
use PUBLIC statements instead.

Change-Id: I18774c95eb6ff57c61212fd3ee2322ba9cf88c3a