AArch64: Fix FPU handling with virtualization
In the case of a virtualizaion enabled kernel we have to take care of the
CPACR_EL1 register when dealing with extended vCPU threads. The current
handling does correctly set the CPACR_EL1 for all normally executing
threads to 3 << 20 to enable the normal in-kernel FPU trapping. However,
there used to be a leftover in the FPU handling that incorrectly saved
and restored the CPACR_EL1 register which is removed by this change.

Fixes ticket #887.

Change-Id: I7575cac7366ba0fbcd8bff855ac3ad3cdb7bbbde