Skip L1D_FLUSH when the flush is not necessary
When the SKIP_L1DFL_VMENTRY bit is set in the IA32_ARCH_CAPABILITIES
MSR, the L1D_FLUSH before entering a VM can be skipped. This is used in
nested virtualization scenarios when the first-level VMM is supposed to
do the cache flushing. Also this bit can be used to signal that the CPU
is not vulnerable to the L1TF-based attacks which would also render the
flush unnecessary.

Change-Id: Ife3679e1bb38fec46ff48a8d03a0511215fe3bef