x86: silence clang warning
Since version 4.0 clang warns about taking addresses of members of
packed structures. The result may be an unaligned pointer value. While
this may cause problems on some architectures it doesn't on x86.

So we disable this warning for ia32 and amd64 when the compiler is clang
and its version is greater or equal 4.0.

Change-Id: I3d0388b72776270735d8274a5d03c2d122c82f1f