aarch64: add support for gcov
This change introduces support for coverage collection using
the gcc feature 'gcov'. It is currently only enabled for
aarch64.

Coverage information is written to the serial line when either
- a systemcall is issued on the Jdb object
- the jdb command '&' is issued

*NOTE* if CONFIG_GCOV is set, there will be a Jdb kernel
object regardless if CONFIG_JDB is set. If CONFIG_JDB is not
set, it can only respond to the gcov call.

Change-Id: Ie7ac05bddf929ac8d5b2e17751f8db9095c1df09