ia32: assert during link on kernel image too big
Depending on the enabled debugging features the kernel can grow beyond
Config::SUPERPAGE_SIZE, in particular on AMD64 with 2MB super pages.

Also increase the kernel map area from 2MB to 4MB on AMD64.

Add an appropriate linker assertion to catch this case. Without this
assertion the code for setting up the initial mappings doesn't add
mappings for initcall data resulting in a screwed up initial IDT. And
of course many other funny things can happen if parts from the tail of
the kernel image are not mapped.

Change-Id: Ie1a2339087277d16b6b7977cb2708dca4227c63a