MIPS: disable generation of small data section
We don't offer space for .sdata in our linker script but Clang generates
a few entries for that section by default leading to linker errors. At
the moment this parameter applies only to Clang (where it is required)
but it might make sense to apply it for GCC as well (in the future).

Change-Id: I746a7b76a6260515ec84729e635827066f5bed6e