Remove cap_has_child function from task protocol
The function to query if a Kobject is still referenced is no longer
needed in userland and is removed from the kernel and the userland
api for the task protocol.

To test wether an object you have handed out is still referenced by
someone else retain a capability to that object and decrease the
reference counter. This effectively leads to a weak reference to the
object. Then call validate() on the capability to check if the object
still exists.

Change-Id: I66acd8219c618205139383a4d83c2c18d8970227
Requires-Change-Id: I8daccd7eb1d6aa8717b95df3a76f2848ca1fb1d8
Requires-Change-Id: I0cb3297f299116004a6b89256f45bf044e9a5526