Remove traps from user-level to kernel debugger
In particular on x86 we used to trap int3 and int1 exceptions from
user-level to the kernel debugger. This change removes the code to do
this completely and always reflects traps to user-level either as
exception IPC or as vCPU upcall.

This also removes the use of F_debug_exc from the vCPU flags.

Change-Id: Ifc06dd5bd223d34921f7788ef8c372ac4250a0fc