forbid unaligned access for capstone contrib code
Fixes the disassembler for ARM.

Gcc generates code for unaligned memory access for certain CPU types.
The capstone library contains code like

    uint8_t *Ptr;
    ...
    NumToSkip = *Ptr++;
    NumToSkip |= (*Ptr++) << 8;

which sometimes is optimized to a single unaligned 16-bit access. This
is known to happen in ARMGenDisassemblerTables.inc which is included by
ARMDisassembler.c but play safe and add -mno-unaligned-access for every
module of that library.

The alignment check is always enabled for 32-bit ARM with virt+lpae. For
64-bit ARM the alignment check is enabled if CONFIG_ARM_ALIGNMENT_CHECK
is enabled but unfortunately this switch doesn't work for AArch64. Maybe
it's not necessary there.

Also do not pack structures for ARM (we already prevent that for MIPS).

Change-Id: Ibf5e97c20dbba5f4b5e0a38702440bcd33821f64