jdb/arm: Use kernel mapping only if it is writable
In access_mem_task() the existing kernel mapping is preferred only if it
is writable. Otherwise the function creates a temporary mapping. This
might be useful for creating writable mappings of kernel memory that is
not writable. When the temporary mapping does not need to be writable,
the existing kernel mapping is still preferred.

Change-Id: Ia39295f9ca7bdf393347d8d7bd0beeb1d64e2da9