fix killing of a thread which never ran before
If a thread never ran then its stack pointer still points to the
following context. Consider this by calling an Assembler function which
sets up a dummy Return_frame before the actual kill function is called.
This is necessary to be on the safe side if the kill function or any of
its callees read or modify a member of Return_frame.

Fixes #CD-121.

Change-Id: I2cd630c262bbacafa53a5f0973eb0e9c045aca63