amd64: use correct type in PCID allocator
When PCID is turned on it causes e.g. spurious segfaults
when L4Linux is under load and creating many address
spaces. This is because Asid_alloc derives from Id_alloc
and instantiates it with unsigned char. It then passes a
constant (1 << 12) - 1 to its constructor.

Change unsigned char to Unsigned16 in the instantiation
of Id_alloc.

Fixes ticket #CD-174.

Change-Id: If37e9d0745bc55fc20b9a45a4ca41579417de6a0