ARM32: force 8-byte AP stack alignment
The stack of an application CPU has to be 8-byte aligned as well,
otherwise long long values are not properly interpreted with va_arg,
for instance in printf().

Note that the 'adr' instruction will only accept certain constant
values, therefore it's not possible to just force the alignment of
the stack. For example, loading '0x400' works while loading '0x404'
does not work.

Change-Id: Ibad1a3d6a456bfaad16cdef4d604660658bdca54