AArch64: Use syscall indirection via the KIP
This change adds a kernel feature to AArch64 similar to
AArch32, where the kernel provides information about how to
do a system call.  On AArch32 the system call code is
provided at fixed location in each address space. However,
on AArch64 this is not easily possible because 64bit user
address spaces are limited in size, depending on the
execution mode and hardware features. To solve this problem
the kernel provides information to enter the kernel in two
ways (a) in the KIP at offset 0x800 and (b) in register x1
on thread startup.

L4Re uses this new information to do system calls. Sigma0
and Moe use the information in x1 because they have no fixed
location of the KIP and Moe cannot find the KIP without
doing a system call already.  All other L4Re programs use a
link-time fix KIP location and do a call via the KIP.

Change-Id: I9f6e800f46d33c725a4ca0365a19e5a5c2b26447