jdb_list: fix seeking in filtered list
If the list is filtered, seeking is done twofold: First, seek the list
starting from the current element to the next element (either forwards
or backwards). The seek() function changes the item if seeking was
successful, that is, if we moved for at least one item.

As the second part we still need to check if the new element is visible
with the filter applied. If not, try to seek again. If yes, check if we
did the requested number of steps and, if so, return the most recent
element we found. Otherwise remember the current item as valid item.

The important thing happens if the seek() function returned '0' which
means that no further step could be done. In that case we have to return
the last element we confirmed to be valid for the filtered list -- not
the element which was just returned by seek()!

Test case: Open the kernel object list ('O') and set a filter to a
string which is not contained in the first element of the list. Then
press the <HOME> key for going to the beginning of the list and then the
<END> key for going to the end of the list. Without this fix the cursor
is now off-by-one beyond the last item.

Change-Id: Ia15aab52cc2cd2accb7d599dda9fc487e6af642a