ARM32: fix clobbers in switch_gp_regs
We have to clobber really all GP registers except r11/fp
which is explicitly saved/restored as part of the
Assembler code.

Without this change the compiler might use 'fp' as one of
the four arguments of the inline asm, and this might lead
to invalid contents in r2 or r3, because one of them would
not be saved correctly.

This resulted in invalid code under certain conditions
(NDEBUG enabled, JDB_LOGGING disabled): The register r3 was
was used as'this' pointer for Context::handle_drq() which
is called right after switch_gp_registers().

We now explicitly use r2/r3 as input/output registers,
so the compiler generates code to handle r2 and r3
correctly.


Fixes ticket #CD-326.

Change-Id: Idb847b1931a2c8dbfc0b81b528f1b86da5388097