AArch64: proper align Map_base for non-virt
This is required for platforms where RAM_PHYS_BASE is not 1GB-aligned.

Without this change, the kernel tries to map unaligned physical memory
to the 1GB-aligned Map_base. This prevents other RAM regions having a
proper 1GB alignment from being mapped with 1GB PT entries.

In other words: Using the same (mis-)alignment for Map_base as for
RAM_PHYS_BASE allows to map RAM regions with a proper GB size and a
proper GB alignment to be mapped with 1GB PT entries.

Related to ticket #CD-108.

Change-Id: I6c66785095a33691ac603f2f2c73dbe206ca0b63