more reliable timer loop calibration
The printf code disables interrupts as it grabs the Libc backend printf
lock. After returning from printf this could lead to a burst of timer
interrupts. As a result, the measurement intervals for Delay::measure()
would be far too short leading to wrong Delay::count values.

So don't block interrupts between enabling the timer and the end of the
measurement.

See also ticket #CD-341.

Change-Id: I9bd742c8f69b9810a25724dd9f3374dfbaa05e71