MDB: Add Mapping_tree::Iterator
This change adds a Mapping_tree::Iterator class to allow
iterating over the mappings for a single physical memory
frame in a consistent way. The iterator is a special
iterator that can be checked for validity by
'*i != nullptr', this allows us to check for the end of
iteration  without having an 'end' iterator to compare
against.

Change-Id: Ib10694f2196d712a28a418e12aad4ad5b6ddb03d