MDB: Avoid using parent is save_access_flags
This removes the case where we stored access flags at the parent
node when doing an unmap including ourselves...
The original implementation is strange anyways, because it was not
possible for the parent itself to read those propagated access flags
without unmapping it's own mapping.
So in order to throw away the parent link in the mapping nodes
we drop this feature here.

Change-Id: I5588c4c56118e8faf73d3b9a8b9d5ed3c96a4587