MDB: rework upgrade/overmap logic
This change provides a new implementation for the overmap
and upgrade logic. The new version avoid multiple searches
in the mapping database but uses a single search run to find
all possible mappings in the case of an overmap. This also avoids
the use of the Mapping::parent() which could be quite expensive.

Change-Id: Id1246d3c03e21e600217691c1f2e5ea9e92d3c7c