Arm/Virt: Allow for more than 16 cores
Spread out the MPID to the next cluster when going beyond 16 cores.

To start QEMU with more than 8 cores, the GICv3 emulation has to be
used. In that case QEMU uses a cluster size of 16 (4 bits) and uses
affinity level 0 of the MIPDR register to distinguish between CPUs
in the same cluster. The remaining bits of the CPU ID are used for
affinity level 1 occupying bits 8..15 of the MPIDR register.

Shift bits 4..7 of the CPU number to bit 8..11 when booting a CPU.
This allows to boot up to 256 CPUs on QEMU/virt.

Change-Id: I61992376c2972086052b150f802c87978de56ef0