Fix deletion for dead and bound thread
If a thread has been bound to a task but
not started yet, call unbind to free
the task's reference. If the thread
would have been started it would terminate
itself and call unbind meanwhile.
Also de-inline the function as we're
calling from more / multiple places.

This addresses ticket #CD-127

Change-Id: I3bda60caf6a38a00e40dec4958b57023986c0d8c