vcpu: treat vcpu with Vcpu_state::F_irqs as open wait
Commit 0c69a8e5 introduced a regression so that a thread in vCPU mode
was not treated as open wait anymore. The underlying problem is that
Receiver::_partner is set already in Receiver::vcpu_async_ipc() which
leads the later code to assume that the receiver is actually in a
closed wait.

This special case is already properly reported by
Receiver::sender_ok(). Just forward the result to the underlying call
sites.

Fixes ticket #CD-371.

Change-Id: I60ab60efd0c6e98bdcc20a17755038ee835f4146