Check _dir was allocated before dereferencing it
Obj_space_phys can get constructed with a NULL _dir and caps_alloc()
should take that into account. This does not appear to be a serious
issue at the moment because the mem variable needs the same amount of
quota as _dir, so we can assume that if _dir allocation failed, so will
that of mem. If mem is not allocated, _dir will not be dereferenced.

However, if _dir allocation size changes in the future (e.g. gets twice
as big) then the dereference will become reachable.

Change-Id: I2eb342df0ada87435e8a6530a4d911ab736a3966