abort IPC during send if the receiver set an error
If the sender was waiting for the receiver in 'Thread::do_send_wait()',
the IPC transfer is performed by 'Thread::ipc_send_msg' in the context
of the receiver. If the IPC fails here, the receiver sets the IPC error
code of the sender.

So far the 'Thread_transfer_failed' state flag signaled that the thread
was killed in the middle of an IPC. Now, re-use that flag to signal that
the IPC failed due to a failed map operation.

The respective UTCB error is now set together with setting the
'Thread_transfer_failed' flag:
- In 'Thread::ipc_receiver_aborted()' if the thread was killed (old
  meaning of 'Thread_cancel/Thread_transfer_failed').
- In the error branch of 'Thread::ipc_send_msg()' if the IPC transfer
  was not successful ('Thread::transfer_msg()' returns false).

The UTCB error code is no longer set in 'Thread::do_send_wait()'.

Fixes ticket #CD-414.

Change-Id: I74c9f1c8a4607f369e0f00b9138b8fae5e769346