add 'S' right check to Semaphore::sys_down()
L4::Semaphore::down() is essentially comparable
to L4::Rcv_endpoint::bind_thread, in the sense that multiple
clients of a service might share the same IRQ or Semaphore
and are allowed to call L4::Triggerable::trigger() (-> Semaphore::up)
but are not allowed to do bind_thread or semaphore down.

This change adds a check for the 'S' right to semaphore down.

This addresses ticket #CD-422.

Change-Id: I7abc8e1bfbda08b652233a1e1c2604b1f63d8810