utest: introduce utest feature string
The feature string acts as placeholder to allow replacing it with
information present externally.
The initial information set contains command line parameters and the
number of cores of the current platform.

The information is encoded in strings and parsed during test startup.

As a first step, the Utest_debug::printf function check the verbosity
parameter provided via the external configuration.

Change-Id: Ib8e9820c7091829f42c006693888ecd8c16b5880