kmem_slab: do not expose possible slab allocator
The generic Kmem_slab_t allocator chooses a suitable allocation back
end based on the object size and possibly other settings. It is thus
not guaranteed that a slab allocator is used. Make sure this is not
visible on the interface and remove the slab references from all call
sites.

Change-Id: I350ec1acc017e53211360f6a88dab562040421b9