minilibc/printf: fix counting of max characters
Specifying the precision of a string allows to limit the maximum number
of characters to be printed. The minilibc variant of printf() checks the
provided string for escape sequences. The check assumes that the string
is terminated by '\0' but it is valid to call printf() with a "%*.s"
format specifier with a non-terminated C string.

Change-Id: Ib383aed1dd9f51189738a678c8446a11f9ddb920