smc_user: add missing fast call check
The conditional 'if ((r0 & 0x3F000000) < 0x30000000)' in kinvoke()
only makes sense if the function identifier is specified using a
fast call (i.e. bit 31 of r0 is set) . This filter doesn't work as
expected if the user performs a yielding call, because the service
call ranges for Trusted Application Calls and Trusted OS Calls are
differ depending on the type of the call (fast or yielding call).

The added check blocks a caller attempting to perform a yielding
call, returning L4_err::ENosys as error code.

Change-Id: Ia5ba1831fe2d11dc34779c65c69683bb9ac04ee7