Fit_size: Restructure and tolerate unordered adding
The Fit_size class of Mem_space assumes that for any given entry in the
array the returned value is the largest valid page_order that is less
than or equal to the index.

The Fit_size is constructed from a size array that needs to also have
the aforementioned property. Adding to this array is done in each type
of Mem_space separately duplicating the code to maintain the property.
Also, when orders are added in a non-increasing fashion, smaller orders
at the moment may overwrite previously added larger orders.

This change encapsulates the logic of adding to the array in Mem_space
into the Fit_size class avoiding duplication of the code. Further we
also ensure that, if a smaller order is added after a larger one, the
smaller order does not overwrite valid larger orders for larger sizes.

Change-Id: I992be8b1974e2ad1cdd365eb8a96f2cb9a2bc44e