encapsulate access to Kip::clock into a function
Access the 64-bit value on 32-bit hosts is not guaranteed to be atomic.
Hence, don't access 'Kip::clock' directly but use 'Kip::clock()' and
rename the member.

This change is also a preparation for a future (configurable) change
when the KIP clock and the system clock are no longer a counter but
rather are synchronized with the scheduling timer. Hence, reading the
system clock means reading the scheduling timer.

Change-Id: I642b5ada8955b75aaec5709d36784441cba059b7