more reliable invalidation of the reply capability
The client is responsible for invalidating the reply capability (caller
ID) of the server. Reset the caller ID before the send operation starts
because at this time the caller ID was already read in case of a reply
operation. The client code for invalidating the caller ID might run too
late.

See ticket #CD-437.

Change-Id: I8f58ec8e463bd103bfdd3bc3cfc4758ae4bbdb71