arm: Add MSI support to GICv3
The GICv3 architecture provides support for message-based interrupts,
e.g. Message Signaled Interrupts (MSI), with the help of the Interrupt
Translation Service (ITS) component.

The kernel configures the ITS to map the combination of a DeviceID and
an EventID to a Locality-specific Peripheral Interrupt (LPI) directed
to a redistributor (i.e. to a CPU). To trigger an LPI, a device has to
write the corresponding EventID to the GITS_TRANSLATER register of the
ITS. The DeviceID for a write, which is unique for each device, is
presented to the ITS in a way that cannot be spoofed. Thus, the ITS can
ensure that a device can only trigger the LPIs that have been mapped
for it.

So far this code has been tested on Arm's AEM FVP simulator and on
the Neoverse N1 hardware platform.

Change-Id: I2013484a3fd9ebe72289d29b088ab784e306789f