Use utility function for icu_helper checks
The checks for the protocol are already implemented in the check_basics
helper function. Use it to simplify the code.

In addition this fixes that the first word in the UTCB might have been
accessed even though the message was sent with no words, since there was
no check for the word count prior to accesing the first element.

Change-Id: I69b7713e56fade4412d64e03ccf969c9b981c011