Cleanup check_basics arguments
The check_basics function performs a rights check that will always
succeed if the (optional) needs parameter is not given. However, the
rights parameter that the needed rights are checked against must always
be provided even if no checking is done due to a missing needs argument.

This change provides an overloaded variant of check_basics which does
not require the rights parameter if no checking is needed.

Change-Id: I6574f78aa09524d5218ec54ac783887f68327b56