Fix potential nullptr deref in Irq_sender
If the IRQ handler thread is on a different CPU than where the IRQ is
triggered, a DRQ is queued (Irq_sender::handle_remote_hit). The handler
will check if the bound thread is still the same and deliver the
message. Otherwise the DRQ is re-queued but the handler needs to check
too for "invalid" thread case because it might race with
Irq_sender::free() on another CPU.

Fixes ticket #CD-486.

Change-Id: Icb73fc5edcdaedc8b42bbd981eb0322da4c3e007