Irq_base: atomic handling of F_enabled bit
The `F_enabled` bit of `Irq_base::_flags` contains the information if
the corresponding Irq is enabled (`F_enabled` set) or disabled/masked
(`F_enabled` not set).

When masking an Irq, the corresponding `Irq_chip::mask()` method should
be called if the Irq was not masked before. When unmasking an Irq, the
corresponding `Irq_chip::unmask()` method should be called if the Irq
was masked before.

Atomically set the `F_enabled` flag and only call the corresponding
`Irq_chip` method if there was really a state transition. This should
fix a race where on MP systems an Irq is repeatedly triggered on a CPU
core different from the core where that Irq is consumed.

See ticket #CD-473.

Change-Id: I2a5a1470638e37191faa275bedb50aa642cba530