Fix the computation of the next elem on prio list
When removing an element which has the same priority as other elements
on the priority list, is not the head of the list of siblings with the
same priority, but is the tail of the siblings list, the returned next
element cannot be computed by following the removed element's P_list
link (as it is only in the S_list and not in the P_list itself).
Instead, dequeue needs to follow the S_list's head's P_list link. All of
this is already handled by Prio_list::next(), so use that.

Fixes issue #CD-485.

Change-Id: I1c944d7a5214e0f5d44433b086f783d0e61b4b88