Arm/vgic: Add a sync point
Ensure that the VGIC has completed before we switch to user-level.
Practically this is only needed for GICv2 because this one is using
MMIO-based access and needs a DSB, while the GICv3 would need an ISB
which however is handled implicitly by the eret, so the sync function is
empty.

Change-Id: Ic9f304a6832911ad38b02eeb5c83b561db9ce601