IA32: Fix APIC calibration on legacy-free systems
On legacy-free systems, the i8254 PIT is unavailable. APIC calibration
runs into an endless loop. If we are on a recent CPU, where the TSC does
not need to be calibrated, use it instead of the PIT to calibrate the
APIC. The existing way remains as fallback for older CPU generations.

Change-Id: Ib1b73d00e904d297569b919b7b5fbd5ac9ce8544
Signed-off-by: Christian Ludwig <cludwig@genua.de>