preprocess: Add support for nested template classes
Preprocess defines nested classes on the fly when it encounters a
function defintion for a nested class. However, preprocess does not set
a `pretext` attribute for the nested class. For function definitions
with template parameters (class template and/or function template) this
becomes a problem later on, resulting in the following error message:
  > Use of uninitialized value in pattern match (m//) at
  > tool/preprocess/src/preprocess line 1281, <INPUT> line 123.

Here, preprocess tries to determine whether the class for this function
definition is a template using its pretext attribute. But for nested
classes the pretext attribute is not set, hence the `uninitialized
value` error.

When creating the nested class, preprocess has all the information
necessary to determine whether the class is a template. Thus, with this
change, the pretext attribute is now set accordingly.

Change-Id: If26fbca3e1a495a45494d31b1ad8bab9e3c2afa8