Use symbolic constants for HCR_EL2 and SCR_EL3
Use symbolic constants for setting the value of HCR_EL2 and SCR_EL3
registers instead of using literal bits directly. This is a preparation
step for supporting Secure EL2.

This patch does not change any functionality.

Note that Cpu::Hcr_default_bits / HCR_EL2_DEFAULT_BITS is defined only
if !cpu_virt. This might look slightly counter-intuitive, but it is not
the invention of the author of this patch. This simply follows from the
functionality of the implementation. Even if the kernel is compiled
without virtualization features, it still sets up the HCR_EL2 register
when booted from EL3 or EL2.

Change-Id: I7b12adca3a2f6f9589f56d77378d24eed11d60d4