arm: Consider entire INTID range when querying pending interrupt
Previously only the bits [9:0] of the 24-bit INTID (at least on GICv3)
were considered, when determining the INTID of a pending interrupt.
That was sufficient until now, because Fiasco did not use any INTIDs
above 1023. However, this changes with the upcoming support of
Locality-specific Peripheral Interrupts, whose INTIDs start at 8192.

Thus change the mask from 0x3ff to 0xffe3ff, to cover the entire INTID
range. As previously we mask off the bits [12:10] that contain a
source CPU interface identifier for SGIs, which is not relevant for us.

Change-Id: Ib959648d5df796ec12f1ddaf28e73bad8749d70f