arm: fix page attributes of kernel page tables on cpu_virt
In stage2 page tables, the page attributes are stored directly
(MemAttr). A stage2 table is only used when virtualization is enabled,
after the initial PL0/1 lookup of the guest has happened. The kernel,
running in PL2 in this case, will use a stage1 page table, though.
There the page tables store an index into the MAIR (AttrIndex).

Change-Id: I8f7b32d3ed1c563f4069b8dab1a7272f94aa4f76