arm: purge ASID based TLB flush for kernel
When running as hypervisor, the TLB flush instructions for the kernel
are different than those of the user space. Some call sites use the
ASID based functions and pass Mem_unit::Asid_kernel. This does not work
because kernel mappings are not affected by user space TLB maintenance
instructions.

Likewise, some of the call sites are also wrong in the non-hypervisor
case. Because contrary to what the name Asid_kernel suggests, the
kernel uses this ASID only during bootstrap, until the first user space
task is switched in. After this point the current ASID is that of the
active user space task. This implies that TLB maintenance for kernel
mappings must be done for all ASIDs. Additionally, global mappings (as
should be all kernel mappings) are never affected by ASID based TLB
maintenance instructions.

The solution is to get rid of Asid_kernel and convert all the call
sites to use dedicated Mem_unit::kernel_tlb_flush() functions. This
additionally optimizes these callers because only the affected TLB
entry is invalidated.

To make things more explicit, the former dtlb_flush() function is
renamed to kernel_tlb_flush().

Co-authored-by: Georg Kotheimer <georg.kotheimer@kernkonzept.com>
Change-Id: I9bee88bd8c1034a0d49d86ab71e69fc5565ec332