Arm: 32: tramp-mp: Avoid using LPAE mcrr when disabled
The mcrr instruction seems to crash on the Krait CPU in Qualcomm MSM8974
(which does not have LPAE support), which prevents the other CPU cores
from coming up correctly.

Using LPAE or not is a build-time option, so this can be easily worked
around by avoiding the problematic instruction entirely for
configurations where LPAE is disabled.

Change-Id: I734f70a92fd23a0b578b681a08067e0d75905d04