arm64: determine upper address limit at runtime on cpu_virt
When running a cpu_virt kernel on arm64, the user space is getting a
stage2 page table from Fiasco and runs with intermediate physical
addresses (IPA). The IPA size is architecturally constrained to the
implemented physical address size of the CPU. Additionally the
available IPA bits depends on the page table format that Fiasco is
compiled with.

To allow the user space to use the whole IPA address range that is
offered by the hardware the maximum supported user space address is
determined at boot time. This allows us to remove the panic() in case
of 4 level page tables where the hardware implements less bits.

Note that it's still be possible to map something above the announced
upper virtual address space limit if arm_pt_48 is enabled and the
hardware provides less IPA bits. The user space will either get a
"stage 1 level 0 Address size fault" (stage 1 MMU disabled) or a "stage
2 Translation fault" (stage 1 MMU enabled, reflected to the vmm) if a
too large address is accessed.

Change-Id: Ib55221bbdf434d2acf6c15ac47dee6565fbdd76a