Do not use Kobject_dbg at runtime
Calls to Kobject_dbg assume that all CPUs are in a quiescent state.
This is not the case outside of JDB and calls to Kobject_dbg will race
with concurrent Kobject creations or deletions. Locking the _kobjects
list while iterating it is not enough because calls to dbg_id() are
virtually dispatched and the object might be in the process of being
deleted.

Remove all calls to Kobject_dbg at runtime to be on the safe side.

Change-Id: I74b788657d72a5aa40a13573037e62efcf793b07