ia32: request interface for immediate VM exit
Fiasco.OC offers a new interface to request immediate VM exit after a
VM entry instead of checking for pending interrupts on the vCPU.

This interface is necessary because KVM in L4Linux provides means to
guarantee a VM exit following a VM entry.
One use case is to ensure event injection on VM entry and is immediate
followed by an VM exit.
KVM therefore schedules a self-IPI. This IPI sets the pending IRQ flag
in the vCPU state, thus leading to an emulated interrupt injection
failure signalled via the IDT-vectoring info field.
KVM reacts to this by re-injecting the same event and sending the
self-IPI again, leading to a loop.

This new interface removes the need for the self-IPI.

The interface uses bit 30 in the VM-entry interrupt information field.
This bit is marked as reserved in this VMCS field according to the
Intel SDM Vol. 3 Section 24.8.3.

This change is based on a contribution by Christian Ehrhardt, Genua GmbH.

Change-Id: I1f92a7c2a75bec9dadd4965813e0d5d995281a23