Use more obvious variable names and add comments
Improve the readability of the code by using more obvious variable names
and adding comments.

The logic has been also changed very slightly by explicitly outputting
the \r\n combination instead of relying on the fact that the \n after
the \r is contained in the next part of the string to be processed.

Change-Id: Ia5e3c7703e693ed26f65271a5126265b228c38fc