Allow disabling of slab caches
Using slab caches greatly reduces allocation overhead at the expense of
slightly more memory overhead. Add an option to disable the usage of
slabs and instead use the boot-time list based best-fit allocator. This
is much slower but will minimize the allocation overhead of kernel
objects as much as possible.

Change-Id: Ib376c6546db8f4da822e2c15926f6d5513e20a19