use -fno-builtin to prevent forbidden optimization
The option -fno-builtin is supported by all relevant gcc and Clang
versions. Without this option, the compiler is free to replace arbitrary
code by a call to an external function providing the same semantics. For
instance, instructions to copy a number of bytes could be (1) replaced
by a few instructions, or (2) by a call to memcpy(), depending on the
available information at build time.

(1) and (2) are desired in certain cases but (2) it is explicitly
forbidden for the _implementation_ of builtin functions: Code in
memcpy() must not be replaced by a call to memcpy() because that would
result in a recursive function call.

Gcc version 12 now replaces the code in strlen() by a call to strlen()
Disable the automatic function replacement for implementation of the
minilibc.

An alternative approach would be to mark respective function
implementations with

  #pragma GCC optimize ("no-tree-loop-distribute-patterns")

This would be more fine-grained but this pragma is not supported by
Clang and the gcc documentation of that option does not describe the
expected behavior. The behavior is only mentioned in gcc bug 56888.

Also remove the global -fno-builtin option for Clang to have the same
behavior with both compilers.

Change-Id: Ib862e647dea705a6e8d57377e8074b309c3f25d1