simplify make rules
Instead of creating a dedicated set of rules for dedicated targets, a
single rule is used for a specific source/destination pattern, composing
the command line according to the module's membership in a specific set
of files.

This makes the used options more obvious and also makes it easier to add
more rules.

Change-Id: I0a665562900103351e169a9e0d40fe4c4b39b762