Cpu_mask: Add atomic_get_and_set_if_unset
Add a method atomic_get_and_set_if_unset that has the semantic of only
writing if the bit is actually unset, returning the previous value of
the bit.

This method is intended to avoid exclusive cache lines being bounced
back and forth between cores. Moreover, the returned information can
be used to perform certain expensive synchronization operations with
other CPUs only if the bit was changed.

Change-Id: I22852646b76960590aef0b0199057136c0d558d7