Kconfig: enable HAS_FPU_OPTION for x86 and amd64
Selecting this option doesn't hurt because both platforms force-enable
the FPU anyway. But it makes a difference if the configuration starts
from x86 or amd64 and then the platform is changed to MIPS.

Before this change, FPU=n was recorded internally because this option
was not set but FPU was forced to be enabled due to the platform. When
changing the platform, FPU=n defines the default if FPU is not forced
to 'y' (MIPS case).

On MIPS, having the FPU enabled is usually desired because the compiler
generates FPU instructions automatically. It should be still allowed to
set FPU=n.

Using this patch, we start with FPU=y when coming from x86 or amd64 and
hence don't override the MIPS default FPU=y.

Fixes ticket #CD-239.

Change-Id: Id0755da8ce2bc08f794b3b5ebb28cf66865e2fc8