arm: support direct vIRQ injection into vCPU
An Irq_sender can be bound to the extended vCPU user thread as virtual
IRQ instead of the regular thread (or vcpu kernel thread) as IPC
destination. The vIRQ will be directly injected into the guest. If the
thread is not in vCPU user mode an optional doorbell Irq is triggered
to let the VMM know that a vIRQ was injected and that the guest should
resume execution.

The VMM should leave some LRs in the vgic empty for usage of the
kernel. Conversely the VMM must be prepared to find populated LRs for
the bound vIRQs that the VMM has not created.

Change-Id: I1ddc25bb430f2fae7467e228e6ed0a73f74d8101