arm: implement IRQ priorities
The priority of an Irq_sender can be set by user space. It will default
to the priority of the thread that is first bound to the Irq_sender. The
priority may be lowered but not increased above the priority of the
currently bound thread. The priority of the Irq_sender is also the one
that is used when queuing it in the sender list. This has the effect
that IRQs will be received according to their priority if multiple IRQs
are pending.

The priority of a hardware IRQ is set in hardware. The priority of the
running thread is also set in the GIC as priority mask. Only IRQs that
have the same or higher priority as the currently running thread are
allowed. This effectively blocks interrupts that would not be handled
anyway.

In support for virtual machines the extended vCPU user state can have a
lower IRQ priority than the normal state. The IRQ priority is derived
from the pending vIRQs with the help of a user space defined linear
function. The priority of the vtimer IRQ can be adjusted in the
extended vCPU state because it has no Irq_sender object.

Change-Id: Ibb1ba144578c79e785dfd6ed243d7cbe8d2bfd7b