arm32: Forward user mode debug exception to user mode
Debug events are reported differently between AArch32 and AArch64. In
AArch32 state a data or prefetch abort with a "debug event" is reported
and the details are encoded in a different register (DBGDSCR.MOE). In
contrast, on AArch64 there are dedicated ESR.EC values for the
different debug events. Because we need to report the root cause to the
exception handler, we convert the syndrome to the AArch64 format.

This has the additional benefit to hide a Qemu bug. Qemu will
incorrectly report the trap of a bkpt instruction to EL2 as HSR.EC ==
0x38. Now that we report the correct event (prefetch abort, IFSC ==
0x22, DBGDSCR.MOE == 3) with the same syndrome to user space there is
no observable difference in behaviour from user space point of view.

Co-authored-by: Jan Klötzke <jan.kloetzke@kernkonzept.com>
Change-Id: I429057ea352f50181a3da9813b485b82b4ac14e8