Fix race conditions in IPC gate del IRQ
The handling of registering and deregistering of IPC gate deletion
IRQs could lead to inconsistent IRQ <-> Thread references in the
kernel. This change fixes the behavior by prohibiting to bind a
new IRQ to the IPC gate deletion signal of a thread if there is
already another IRQ object bound. Additionally this change uses
the irq_lock of the IRQ object to correctly protect bind and unbind
operations. As well as it uses mp_cas to consistently handle
writes to Thread::_del_observer.

This change is related to ticket #CD-676.

Change-Id: I778754ea1a566cdf4cdd3adb8437e3f7a072f2c2