arm: Ensure current tpidrurw is used for exception IPC
Previously, during exception IPC it could happen that instead of the
current value of a thread's TPIDRURW register an outdated value, stored
in Context::_tpidrurw, was transferred to the exception handler.
Then when processing the exception handler's reply, the current value
gets overwritten with the outdated value from the response.

Specifically, I observed this behavior with a vCPU, where the bug
resulted in a page fault in host mode, because after an exception
IPC its TPIDRURW register was set to the value used by the guest.
However, also regular threads can be affected by this, although it is
less likely, since they usually set their TPIDRURW only once initially,
and don't change it afterwards.

Change-Id: I1b825e2450e7652e64c4755d5d53033d3f35c98d