arm: map pages eager in standard tasks on MPU systems
On MMU systems, we rely on demand paging for sigma0 and the root task.
On MPU platforms it is essential to have contiguous regions in tasks,
though. Such platforms provide only a limited number of MPU regions. On
the other hand we cannot simply map too much because MPU regions must
not overlap with the kernel. Instead, let bootstrap mark precisely the
regions that shall be mapped eagerly.

Change-Id: If0cd7efe04c2e0eaabe9a2e52c2e0a2ec8b0ef9b