arm: add MPU platform support
All paging data structures are removed and replaced by MPU region
lists. The region list is independent of the actual hardware and also
used if no MPU is used at all. The region list is then synced to HW and
user space regions are activated kernel exit.

Because MPU regions must not overlap it is not easily possible to map
all UTCBs in the kernel and only selected ones when in user space. This
is solved by switching the active MPU regions on kernel entry/exit.
This needs to be optimized in the future to switch only what is
absolutely necessary. It might also be possible to do it lazily.

To support virtualisation the extended vCPU state needs to hold the EL1
MPU settings. This is saved in a union to reuse the MMU register space
that is not needed on MPU platforms.

Change-Id: Ia5fec99d2527544ba0c6a3e2815d766c2b5da914