Arm: bsp: qcom: Add IOMMUs (SMMUv2)
Most modern Qualcomm platforms have two IOMMUs (standard ARM SMMUv2):

  - APPS SMMU that covers most DMA-capable components in the SoC
  - GPU SMMU (probably with special optimizations for performance)

The memory addresses and interrupt numbers vary even between many minor
SoC versions (such as MSM8909, MSM8916 and MSM8939 which are very
similar otherwise). Adding support for the IOMMU mainly involves big
blocks with the correct definitions for the different platforms.

IOMMUs are also present on MSM8226 and MSM8974 but not added here
because the setup is a bit different. Also note that most firmware
versions "mess up" the SMMUs by locking parts of the register space.
The IOMMU option can be only used if the SMMUs are exposed to Fiasco
in a standard way.

Change-Id: I90dd80227217dc213714dd68d02af6e21f51068f