jdb: simplify waiting for cursor position report
The function 'wait_for_escape_sequence()' is exclusively used for
waiting for the 'ESC[#;#R' cursor position report after 'ESC[6n' was
sent to the console.

So far we used the RDTSC instruction on x86/amd64 and a loop with a
fixed loop count on other architectures to wait for a certain amount
of time while detecting the escape sequence.

On newer ARM processors, the loop count was too small. Unify the code
and use Delay::delay(1) as a reasonable accurate delay of 1ms on all
architectures and wait for no more than 100ms.

Change-Id: Ief8d1b9b294e7091632fac25e54d00de01590046