riscv: Always inline methods used by bootstrap code
We have to ensure that methods that are used in bootstrap code are
always inlinied, i.e. also when building Fiasco with the
"Generate inline code" option disabled. This is important, because
otherwise linking the bootstrap code fails with linker errors.

After the last rework of the bootstrap code, it was forgotten to
update the methods that always have to be inlined.

Change-Id: I209b78367cdcbacabfe5bf57dcee22259bf17f1c