arm: allow selective initialization of gic distributor
If multiple nodes share the same GIC distributor, only the first node
must initialize it. Add a parameter to the Gic_v3 constructor to allow
skipping initialization on secondary nodes.

So far, the Gic_mixin::init() method was never called with 'false' as
first parameter. That allows to repurpose this parameter for selective
initialization.

Change-Id: Ic9c14aafe87739eb83b62a50bac30024dee6863e