Flag empty maps in received message item.
This change adds bit 1 (0x02) into the send-base of a received message
item if the message item was a map item that resulted in a actual map
operation and this map operation returned successful but did not map
anything. This is in particular defined in the case of object maps and
signals the receiver that it has not received any caps in its receive
buffer.

This fixes ticket #CD-730.

Change-Id: I1df905c779eb1a6763077bdc3a9cb5ce4f4330d8
Requires-Change-Id: I45b6ea98055e65cda82d709d078756c5d6f7210c