arm: Move arm32 exclusive code into arm32 source file
Move the default implementation of Cpu::init_mmu to cpu-arm-32.cpp,
because high exception vectors and thus mapping the interrupt vector
table to 0xffff0000 is only a thing on ARM 32-bit. Having ARM 32-bit
exclusive code as a default implementation in cpu-arm.cpp, which is
shared with ARM 64-bit, is rather unintuitive, and in addition requires
an explicit empty IMPLEMENT_OVERRIDE in cpu-arm-64.cpp.

A reversed approach is preferable, with an empty default implementation,
which is overridden with a non-empty implementation where necessary.

Change-Id: I9575b95ddeb7b5d556608bfc87a139cd63a7d81c