Add notion of multiple nodes
The KIP is reorganized to make room for up to 8 nodes. For each node
the start address of sigma0 and the root task can be specified
independently. The unused stack pointer fields are dropped to make
room. This breaks the user space ABI.

The platform control object knows the node id. When the initial tasks
are started this node id is taken into account.

Change-Id: I6b37a5a0991ab886adbbbed2ceda4147ea1d4802